LayoutTests/platform/ios-sim-deprecated/accessibility/password-value-expected.txt


This test that a password field returns a string of equal length to the password, but not the actual password

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pass.iphoneValue.length is password.length + axValue.length
PASS pass.iphoneValue == password is false
PASS successfullyParsed is true

TEST COMPLETE
